services/python/requirements.txt
fastapi[standard]>=0.114.0,<0.115.0
pydantic>=2.7.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0